Initial version RunKeyLauncher
RunKeyLauncher will run an AHK script you are editing in VSCode, under
the condition that the script in question is stored in a place that RKL
knows about. You can add paths, one per line, in the RKL-SELFCONFIG file
that is generated the first time a script runs (or when it runs and
cannot find that file.) This will be updated in future with more and
better features. It is bound to the F5 key, which is "run with debugging
" in VSCode.